backend/requirements.txt
asgiref==3.5.0
astroid==2.11.5
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dill==0.3.5.1
dj-rest-auth @ git+https://github.com/iMerica/dj-rest-auth.git@bf168d9830ca2e6fde56f83f46fe48ab0adc8877
Django==3.2.14
django-allauth==0.51.0
django-cors-headers==3.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
drps==1.2
flake8==4.0.1
h11==0.13.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kwonly-args==1.0.10
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
pwny==0.1.5
pwnypack==0.9.0
py-cpuinfo==8.0.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pylint==2.13.9
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.10.0
typing-extensions==4.2.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.14.1
wsproto==1.1.0
